Refactor: API filenames to include HTTP method prefix
