+++ api/src/routes/employees/delete_delete.ts
